Moved MSBuild tests
Moved the MSBuild tests into the appropriate DLL
